# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/user":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["User"],"security":[{"JWT":[]}]}},"/user/{id}":{"get":{"operationId":"UserController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]},"put":{"operationId":"UserController_update","summary":"Create a new user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]},"delete":{"operationId":"UserController_remove","summary":"Delete an user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]}},"/location":{"get":{"operationId":"LocationController_findAll","summary":"Get all locations","parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Location"}}}}}},"tags":["Location"]},"post":{"operationId":"LocationController_create","summary":"Create a new location","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}}},"tags":["Location"]}},"/location/{id}":{"get":{"operationId":"LocationController_findById","summary":"Get a location by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}}},"tags":["Location"]},"put":{"operationId":"LocationController_update","summary":"Create a new location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}}},"tags":["Location"]},"delete":{"operationId":"LocationController_remove","summary":"Delete a location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}}},"tags":["Location"]}},"/store":{"get":{"operationId":"StoreController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Store"]},"post":{"operationId":"StoreController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreDto"}}}},"responses":{"201":{"description":""}},"tags":["Store"]}},"/store/{id}":{"get":{"operationId":"StoreController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Store"]},"put":{"operationId":"StoreController_update","summary":"Create a new location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Store"}}}}},"tags":["Store"]},"delete":{"operationId":"StoreController_remove","summary":"Delete a location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Store"}}}}},"tags":["Store"]}},"/local/register":{"post":{"operationId":"AuthController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"201":{"description":""}}}},"/local/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}}}}},"info":{"title":"Ludu API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UserDto":{"type":"object","properties":{"role":{"example":"Location 1","description":"The name of the store","type":"array","items":{"type":"string"}},"username":{"type":"string","example":"popolito","description":"Your nickname"},"credentials":{"type":"object"},"phone":{"type":"string","example":"User's phone number","description":"+33620202020"},"address":{"type":"string","example":"User's adresse","description":"16 rue de beaumont, Montesson 78360"},"stores":{"description":"Stores","type":"array","items":{"type":"string"}}},"required":["role","username","credentials","phone","address"]},"User":{"type":"object","properties":{}},"Location":{"type":"object","properties":{}},"LocationDto":{"type":"object","properties":{"name":{"type":"string","example":"Location 1","description":"The name of the location"},"postalCode":{"type":"number","example":"Location 1","description":"The name of the location"},"stores":{"description":"Stores","type":"array","items":{"type":"string"}}},"required":["name","postalCode"]},"StoreDto":{"type":"object","properties":{"name":{"type":"string","example":"Location 1","description":"The name of the store"},"address":{"type":"string","example":"75 rue de test","description":"The address of the store"},"owner":{"type":"string","example":"Santamaria","description":"Name of the shop's owner"},"phone":{"type":"string","example":"+33639393939","description":"Number of the shop's owner"},"location":{"type":"string","example":"Paris","description":"Location of the shop, OneToMany relation"}},"required":["name","address","owner","phone","location"]},"Store":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{}}}}}